requirements.txt
Django==3.0.7
requests==2.24.0
coverage==5.0.4
factory_boy==2.12.0
lxml==4.5.2
beautifulsoup4==4.5.0
django-cors-headers==3.4.0
django-background-tasks==1.2.5